상품 리뷰 및 프로필 페이지 수정

+++ 210731/API/ProductAPI.swift
@@ -11,7 +11,7 @@ import Moya
 enum ProductAPI {
     case totalProducts
     case detailProducts(id: String)
-    case createReview(id: String, token: String, rating: Int, comment: String)
+    case createReview(id: String, token: String, rating: Float, comment: String)
 }
 
 extension ProductAPI : TargetType {
